move estimate settings out into new Estimate class
